There are two bugs, one of which was fixed.
1. the original response was being used to return www-authenticate in
the basic auth implementation.  however, because an exception was
thrown, the header should have been set on the exception (exceptions
act similarly to responses).

2. routing breaks when there are decorators.  this is because the
outermost decorators signature is used instead of the actual method
handler for comparison.  this is a tough one to fix.  here are some
options.

http://stackoverflow.com/questions/14413946/what-exactly-is-contained-wi
thin-a-obj-closure

http://stackoverflow.com/questions/1166118/how-to-strip-decorators-from-
a-function-in-python